+++ packages/manager/apps/web/client/app/hosting/multisite/hosting-multisite.constants.js
@@ -1,0 +1,5 @@
+export const CDN_ACTIVE = 'ACTIVE';
+
+export default {
+  CDN_ACTIVE,
+};