+++ packages/manager/apps/web/client/app/hosting/multisite/statistics/hosting-multisite-statistics.component.js
@@ -1,0 +1,13 @@
+import controller from './hosting-multisite-statistics.controller';
+import template from './hosting-multisite-statistics.html';
+
+export default {
+  bindings: {
+    chartJsFactory: '<',
+    domains: '<',
+    getStatistics: '<',
+    periods: '<',
+  },
+  controller,
+  template,
+};